requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pyyaml>=6.0
